perf(prefetch): server-prefetch + HydrationBoundary for dashboard pages (Tier 1+2) (#893)
* perf(prefetch): wave 1 — org home/analytics + consultee home server-prefetch + dedup

Server-prefetch + HydrationBoundary (extends #669) so data hydrates at first paint:
- org home + analytics via a shared getOrgAnalytics() in lib/data/org-analytics.ts
  (analytics route refactored to use it → one SSR/CSR read path). Key ["org-analytics", orgId].
- consultee home: readConsulteeEvents() extracted (events route refactored) +
  prefetchConsulteeEvents; page split into server + HomePageClient. Key
  ["consultee-events", id, "personal"]. NOTE: org-member default scope is "all"
  (role-dependent, client-only), so org members fall back to a client fetch while
  B2C/personal users hydrate — session-aware scope is a follow-up.
- centralize waitForGlobalVideoClient in lib/stream/disconnect (was duplicated in
  both HomeTabs; resolves the #889 review nits).

Part of #890.


* perf(prefetch): wave 2 — consultant/admin/staff home + standardize convention + RSC-safe

Server-prefetch + HydrationBoundary for the remaining dashboard home pages
(consultant ["consultant-dashboard"], admin ["admin-stats"], staff ["staff-stats"]),
each via a shared read extracted from its route (route refactored to call it →
one SSR/CSR path).

Standardize the convention (one rule):
- every server-side read lives in lib/data/<name>.ts; the server page calls it
  DIRECTLY inside prefetchQuery — no lib/server/*-prefetch indirection.
- moved consultee-events-read into lib/data; extracted org members read to
  lib/data/org-members.ts; deleted lib/server/{org,consultee,admin,staff,consultant}-prefetch.ts
  (lib/server/ is now empty). org home/analytics/members call lib/data reads directly.

Fix (correctness): wrap reads that return Prisma money-$extends rows in toPlain()
(lib/data/serialize) — those rows carry a Symbol that breaks RSC serialization at
HydrationBoundary. Applied to consultee-events-read, admin-stats, staff-stats
(consultant already had it; org-analytics builds a plain object).

Part of #890.


* perf(prefetch): wave 3 — appointment-list pages (consultant/consultee/staff/org)

Server-prefetch + HydrationBoundary for the Tier-2 appointment lists, same
convention (read in lib/data, page prefetches it directly, toPlain on returns):
- consultant: lib/data/consultant-appointments.ts (extracted from /api/slots/appointments,
  all 5 other consumers preserved). Key ["consultant-appointments", id, "personal"].
- consultee: reuses lib/data/consultee-events-read. Key ["consultee-events", id, "personal"].
- staff: lib/data/staff-appointments.ts (page converted from raw fetch → useQuery).
  Key ["staff-appointments", {page:1,...}].
- org: lib/data/org-appointments.ts (wraps listAppointmentsScoped). Key ["org-appointments", id, 1].

Prefetch covers the DEFAULT view only; filtered/paginated/other-scope views fall
back to a client fetch by design (the key includes those). Role-dependent default
scope ("all" for org members/admins) is client-only, so those land on a client
fetch while B2C/personal users hydrate — consistent with the home pages.

Part of #890.


* fix(prefetch): address #893 review — prefetch authz, error states, waitlist isolation, input sanitization, a11y

- Enforce org-role authz before SSR prefetch on analytics/appointments/home/
  members pages (mirrors the API routes: MANAGER for analytics+appointments;
  operator-or-finance surface for members; home guards only the analytics
  prefetch so lower roles keep their overview). The client-only org layout
  provided no server-side gate, so the dehydrated payload leaked otherwise.
- Add explicit error+retry UI to AdminHomePageClient, staff HomePageClient,
  and org AnalyticsPageClient so a failed fetch no longer renders misleading
  zeros / hangs on an infinite skeleton.
- consultee-events-read: apply org scope to the webinar/class waitlist OR
  branch so an org-scoped read can't leak cross-tenant/personal waitlisted
  entries.
- Clamp page/limit to positive integers (staff appointments route + org
  appointments client) — parseInt("abc")=NaN and Number("-1") both bypassed
  the old fallback.
- staff HomePageClient: aria-label/title on the icon-only refresh button.
- consultant HomePageClient: use isFetching (not isLoading) for the
  "Refreshing..." background-refetch badge.


* fix(prefetch): address #893 review round 2 — consultee skeleton flash + org-members payoutRecipient

- consultee appointments: guard skeleton on initial load only so a scope
  switch keeps the keepPreviousData list instead of flashing (#346)
- getOrgMembers: select + expose payoutRecipient so the SSR row matches the
  API shape the edit dialog reads (#729)

org-members SSR prefetch key/shape mismatch tracked separately in #902.

Part of #893


* fix(staff): filter appointments + compute counts in the DB, not per-page

Status was a derived value (per-type sub-entity status + payment presence),
so filtering ran after pagination — page sizes shrank and pagination.total
stayed unfiltered. Tab counts were computed from the current page only.

Resolve the staff display-status (scheduled/completed/issue) to DB conditions
across the four event types and AND it onto the list query, so pagination +
total reflect the filtered set. Compute the tab badges via separate count()
queries over the base filter, so they're global, not page-scoped.

Closes #897


---------
