+++ app/api/admin/stats/route.ts
@@ -1,14 +1,14 @@
 import * as Sentry from "@sentry/nextjs";
 import { NextResponse } from "next/server";
 import { requirePrivilegedAuth } from "@/lib/auth-helpers";
-import { getOperatorDashboardStats } from "@/lib/api/operators";
+import { getAdminStats } from "@/lib/data/admin-stats";
 
 export async function GET() {
   try {
     const auth = await requirePrivilegedAuth();
     if (auth.error) return auth.error;
 
-    const stats = await getOperatorDashboardStats();
+    const stats = await getAdminStats();
     return NextResponse.json(stats);
   } catch (error) {
     console.error("Admin stats error:", error);
