+++ app/api/slots/appointments/route.ts
@@ -1,8 +1,9 @@
 import { NextRequest, NextResponse } from "next/server";
 import prisma from "@/lib/prisma";
-import { AppointmentsType, Prisma, AppointmentStatus } from "@prisma/client";
+import { AppointmentsType, Prisma } from "@prisma/client";
 import { requireApiAuth, isPrivileged } from "@/lib/auth-helpers";
 import { resolveOrgScope } from "@/lib/api/scope/parse";
+import { getConsultantAppointments } from "@/lib/data/consultant-appointments";
 
 export async function GET(request: NextRequest) {
   const authResult = await requireApiAuth();
@@ -137,12 +138,12 @@ export async function GET(request: NextRequest) {
     const startDate = searchParams.get("startDate");
     const endDate = searchParams.get("endDate");
 
-    const appointments = await getAppointments(
-      type as AppointmentsType | undefined,
+    const appointments = await getConsultantAppointments({
+      type: type as AppointmentsType | undefined,
       consultantProfileId,
       consulteeProfileId,
       userId,
-      {
+      statuses: {
         consultation: consultationStatus || undefined,
         subscription: subscriptionStatus || undefined,
         webinar: webinarStatus || undefined,
@@ -150,14 +151,14 @@ export async function GET(request: NextRequest) {
       },
       startDate,
       endDate,
-      {
+      eventIds: {
         webinarId,
         classId,
         consultationId,
         subscriptionId,
       },
-      apptOrgFilter,
-    );
+      orgScopeFilter: apptOrgFilter,
+    });
 
     return NextResponse.json({ data: appointments });
   } catch (error) {
@@ -167,334 +168,6 @@ export async function GET(request: NextRequest) {
       { status: 500 },
     );
   }
-}
-
-async function getAppointments(
-  type?: AppointmentsType,
-  consultantProfileId?: string | null,
-  consulteeProfileId?: string | null,
-  userId?: string | null,
-  statuses?: {
-    consultation?: string;
-    subscription?: string;
-    webinar?: string;
-    class?: string;
-  },
-  startDate?: string | null,
-  endDate?: string | null,
-  eventIds?: {
-    webinarId?: string | null;
-    classId?: string | null;
-    consultationId?: string | null;
-    subscriptionId?: string | null;
-  },
-  /// #674 org-scope filter resolved upstream — { organizationId: null }
-  /// for personal scope, { organizationId: <id> } for an org context,
-  /// {} for admin all-scope. Spread into the appointment WHERE clause.
-  orgScopeFilter: Partial<Prisma.AppointmentWhereInput> = {},
-) {
-  const whereClause: Prisma.AppointmentWhereInput = { ...orgScopeFilter };
-
-  // Date range filtering for appointments. This is the primary filter.
-  // It looks for appointments where any of its slots overlap with the given date range.
-  // Also includes appointments with no slots (they'll be shown but sorted to the end)
-  if (startDate && endDate) {
-    whereClause.OR = [
-      // Appointments with slots overlapping the date range
-      {
-        slotsOfAppointment: {
-          some: {
-            AND: [
-              { startsAt: { lt: new Date(endDate) } },
-              { endsAt: { gt: new Date(startDate) } },
-            ],
-          },
-        },
-      },
-      // OR appointments with no slots at all
-      {
-        slotsOfAppointment: {
-          none: {},
-        },
-      },
-    ];
-  }
-
-  const userFilterClauses: Prisma.AppointmentWhereInput[] = [];
-  if (consultantProfileId) {
-    userFilterClauses.push({
-      OR: [
-        {
-          consultation: {
-            consultationPlan: { consultantProfileId },
-          },
-        },
-        {
-          subscription: {
-            subscriptionPlan: { consultantProfileId },
-          },
-        },
-        {
-          webinar: {
-            webinarPlan: { consultantProfileId },
-          },
-        },
-        {
-          // Collaborated webinars (co-host, moderator, etc.)
-          webinar: {
-            webinarPlan: {
-              collaborators: {
-                some: {
-                  consultantProfileId,
-                  status: "ACCEPTED",
-                },
-              },
-            },
-          },
-        },
-        {
-          class: {
-            classPlan: { consultantProfileId },
-          },
-        },
-        {
-          // Collaborated classes (co-instructor, TA, etc.)
-          class: {
-            classPlan: {
-              collaborators: {
-                some: {
-                  consultantProfileId,
-                  status: "ACCEPTED",
-                },
-              },
-            },
-          },
-        },
-      ],
-    });
-  }
-
-  if (consulteeProfileId) {
-    userFilterClauses.push({
-      OR: [
-        {
-          consultation: {
-            requestedBy: { id: consulteeProfileId },
-          },
-        },
-        {
-          subscription: {
-            requestedBy: { id: consulteeProfileId },
-          },
-        },
-        {
-          slotsOfAppointment: {
-            some: {
-              user: { some: { consulteeProfileId: consulteeProfileId } },
-            },
-          },
-        },
-      ],
-    });
-  }
-
-  if (userId) {
-    userFilterClauses.push({
-      slotsOfAppointment: {
-        some: {
-          user: {
-            some: {
-              id: userId,
-            },
-          },
-        },
-      },
-    });
-  }
-
-  if (userFilterClauses.length > 0) {
-    whereClause.AND = userFilterClauses;
-  }
-
-  if (type) {
-    whereClause.appointmentType = type;
-  }
-
-  // Filter by specific event IDs
-  if (eventIds?.webinarId) {
-    whereClause.webinar = { id: eventIds.webinarId };
-  }
-  if (eventIds?.classId) {
-    whereClause.class = { id: eventIds.classId };
-  }
-  if (eventIds?.consultationId) {
-    whereClause.consultation = { id: eventIds.consultationId };
-  }
-  if (eventIds?.subscriptionId) {
-    whereClause.subscription = { id: eventIds.subscriptionId };
-  }
-
-  // FIX #551: Apply status filters that were previously parsed but never used.
-  // Build an OR clause so appointments matching ANY of the provided status
-  // filters are returned (allows fetching e.g., APPROVED consultations AND
-  // SCHEDULED webinars in one call).
-  const statusFilters: Prisma.AppointmentWhereInput[] = [];
-  if (statuses?.consultation) {
-    statusFilters.push({
-      consultation: { status: statuses.consultation as AppointmentStatus },
-    });
-  }
-  if (statuses?.subscription) {
-    statusFilters.push({
-      subscription: { status: statuses.subscription as AppointmentStatus },
-    });
-  }
-  if (statuses?.webinar) {
-    statusFilters.push({
-      webinar: { status: statuses.webinar as Prisma.EnumWebinarStatusFilter },
-    });
-  }
-  if (statuses?.class) {
-    statusFilters.push({
-      class: { status: statuses.class as Prisma.EnumClassStatusFilter },
-    });
-  }
-  if (statusFilters.length > 0) {
-    // Combine with existing AND clauses
-    const existingAnd = whereClause.AND
-      ? Array.isArray(whereClause.AND)
-        ? whereClause.AND
-        : [whereClause.AND]
-      : [];
-    whereClause.AND = [...existingAnd, { OR: statusFilters }];
-  }
-
-  const appointments = await prisma.appointment.findMany({
-    where: whereClause,
-    include: {
-      slotsOfAppointment: {
-        orderBy: { startsAt: "asc" },
-        include: {
-          user: true,
-          meetingSession: {
-            select: { id: true, endedAt: true },
-          },
-        },
-      },
-      consultation: {
-        include: {
-          consultationPlan: {
-            include: {
-              consultantProfile: {
-                include: {
-                  user: true,
-                },
-              },
-            },
-          },
-          requestedBy: {
-            include: {
-              user: true,
-            },
-          },
-        },
-      },
-      subscription: {
-        select: {
-          id: true,
-          subscriptionPlan: {
-            include: {
-              consultantProfile: {
-                include: {
-                  user: true,
-                },
-              },
-            },
-          },
-          requestedBy: {
-            include: {
-              user: true,
-            },
-          },
-          schedulingPeriodStartsAt: true,
-          schedulingPeriodEndsAt: true,
-          status: true,
-        },
-      },
-      webinar: {
-        include: {
-          webinarPlan: {
-            include: {
-              consultantProfile: {
-                include: {
-                  user: true,
-                },
-              },
-              collaborators: {
-                where: { status: "ACCEPTED" },
-                include: {
-                  consultantProfile: {
-                    include: {
-                      user: {
-                        select: {
-                          id: true,
-                          name: true,
-                          image: true,
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      },
-      class: {
-        include: {
-          classPlan: {
-            include: {
-              consultantProfile: {
-                include: {
-                  user: true,
-                },
-              },
-              collaborators: {
-                where: { status: "ACCEPTED" },
-                include: {
-                  consultantProfile: {
-                    include: {
-                      user: {
-                        select: {
-                          id: true,
-                          name: true,
-                          image: true,
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      },
-    },
-  });
-
-  // Sort appointments by slot start time
-  // Include appointments with 0 slots (push them to the end)
-  return appointments.sort((a, b) => {
-    const aTime = a.slotsOfAppointment?.[0]?.startsAt;
-    const bTime = b.slotsOfAppointment?.[0]?.startsAt;
-
-    // Appointments without slots go to the end
-    if (!aTime && !bTime) return 0;
-    if (!aTime) return 1;
-    if (!bTime) return -1;
-
-    return new Date(aTime).getTime() - new Date(bTime).getTime();
-  });
 }
 
 export async function POST(request: NextRequest) {
