+++ app/api/staff/stats/route.ts
@@ -2,20 +2,21 @@
  * Staff Dashboard Stats API
  * Returns homepage statistics for the staff dashboard.
  *
- * Thin shell — query/aggregation logic lives in
- * `lib/api/operators/stats.ts` (`getStaffDashboardStats`).
+ * Thin shell — the JSON-safe read lives in `lib/data/staff-stats.ts`
+ * (`getStaffStats`), shared with the server prefetch (#890) so SSR
+ * hydration and the client fetch resolve identical payloads.
  */
 
 import { NextResponse } from "next/server";
 import { requirePrivilegedAuth } from "@/lib/auth-helpers";
-import { getStaffDashboardStats } from "@/lib/api/operators";
+import { getStaffStats } from "@/lib/data/staff-stats";
 
 export async function GET() {
   try {
     const auth = await requirePrivilegedAuth();
     if (auth.error) return auth.error;
 
-    const stats = await getStaffDashboardStats();
+    const stats = await getStaffStats();
     return NextResponse.json(stats);
   } catch (error) {
     console.error("Error fetching staff stats:", error);
