+++ app/dashboard/consultee/[consulteeId]/(features)/home/HomeTab.tsx
@@ -26,7 +26,7 @@ import { useState, useMemo, useRef } from "react";
 // lib/meeting (which imports the SDK) here — that would pull the heavy SDK into
 // the dashboard-HOME bundle / critical path. The video client + meeting helper
 // are acquired lazily inside the Join handler (only when a user clicks Join).
-import { getGlobalVideoClient } from "@/lib/stream/disconnect";
+import { waitForGlobalVideoClient } from "@/lib/stream/disconnect";
 import { useToast } from "@/hooks/use-toast";
 import type { TConsulteeEventsResponse } from "@/types/consultee-events";
 import {
@@ -50,26 +50,6 @@ interface HomeTabProps {
   eventsData: TConsulteeEventsResponse;
   isRefreshing?: boolean;
   consulteeId: string;
-}
-
-// #248: the video connect is deferred to requestIdleCallback, so the global
-// client may not exist yet at the instant a user clicks Join. Poll briefly for
-// it (the StreamProvider mounted on this route is already connecting) rather
-// than erroring out on the first null read. Resolves null if it never appears
-// within the window so the caller can show a soft "try again" message.
-async function waitForGlobalVideoClient(
-  timeoutMs = 4000,
-  intervalMs = 150,
-): Promise<ReturnType<typeof getGlobalVideoClient>> {
-  const existing = getGlobalVideoClient();
-  if (existing) return existing;
-  const deadline = Date.now() + timeoutMs;
-  while (Date.now() < deadline) {
-    await new Promise((resolve) => setTimeout(resolve, intervalMs));
-    const client = getGlobalVideoClient();
-    if (client) return client;
-  }
-  return getGlobalVideoClient();
 }
 
 const staggerChildren = {
