+++ app/dashboard/organization/[orgId]/analytics/page.tsx
@@ -1,211 +1,36 @@
-"use client";
+import { HydrationBoundary, QueryClient, dehydrate } from "@tanstack/react-query";
+import { redirect } from "next/navigation";
+import { requireOrgAccess } from "@/lib/auth-helpers";
+import { AnalyticsPageClient } from "./AnalyticsPageClient";
+import { getOrgAnalytics } from "@/lib/data/org-analytics";
 
-import { use } from "react";
-import { useQuery } from "@tanstack/react-query";
-import { useRequireOrgRole } from "../useOrgRole";
-import {
-  Users,
-  Briefcase,
-  Wallet,
-  AlertCircle,
-  FileText,
-  TrendingUp,
-  UserCheck,
-} from "lucide-react";
-import type { FundingSource, MemberRole } from "@prisma/client";
-
-import {
-  DashboardHeader,
-  DashboardContent,
-  DashboardGrid,
-} from "@/components/dashboard/DashboardShell";
-import { StatCard, StatCardSkeleton } from "@/components/dashboard/StatCard";
-import { formatCurrencyAmount } from "@/utils/formatting";
-
-// ---------------------------------------------------------------------------
-// Types — match GET /api/organizations/[orgId]/analytics
-// ---------------------------------------------------------------------------
-
-interface OrgAnalytics {
-  capabilities: {
-    canSponsor: boolean;
-    canHost: boolean;
-    fundingSource: FundingSource | null;
-    walletBalance: number | null;
-    currency: string | null;
-  };
-  members: {
-    total: number;
-    active: number;
-    byRole: Array<{ role: MemberRole; count: number }>;
-  };
-  programs: {
-    total: number;
-    active: number;
-    activeAssignments: number;
-  };
-  wallet: {
-    balancePaise: number;
-    recent: Array<{ reason: string; count: number; deltaPaise: number }>;
-  } | null;
-  invoices: {
-    outstandingCount: number;
-    outstandingPaise: number;
-    pastDueCount: number;
-    paidLast30dCount: number;
-    paidLast30dPaise: number;
-  } | null;
-  earnings: Array<{
-    status: string;
-    count: number;
-    orgSharePaise: number;
-    refundedPaise: number;
-  }> | null;
-}
-
-async function fetchAnalytics(orgId: string): Promise<OrgAnalytics> {
-  const res = await fetch(`/api/organizations/${orgId}/analytics`);
-  if (!res.ok) throw new Error("Failed to load analytics");
-  return res.json();
-}
-
-function countByRole(
-  byRole: OrgAnalytics["members"]["byRole"],
-  role: MemberRole,
-): number {
-  return byRole.find((r) => r.role === role)?.count ?? 0;
-}
-
-export default function OrgAnalyticsPage({
+export default async function OrgAnalyticsPage({
   params,
 }: {
   params: Promise<{ orgId: string }>;
 }) {
-  const { orgId } = use(params);
-  const { allowed } = useRequireOrgRole(orgId, "MANAGER");
-  const { data, isLoading } = useQuery({
-    queryKey: ["org-analytics", orgId],
-    queryFn: () => fetchAnalytics(orgId),
-    enabled: allowed,
-  });
+  const { orgId } = await params;
 
-  if (!allowed) return null;
+  // Mirror GET /api/organizations/[orgId]/analytics (MANAGER) — the SSR
+  // prefetch reads org-scoped analytics directly, so without this the
+  // dehydrated payload would embed manager-only data for any member.
+  const access = await requireOrgAccess(orgId, "MANAGER");
+  if (access.error) redirect(`/dashboard/organization/${orgId}/home`);
 
-  if (isLoading || !data) {
-    return (
-      <>
-        <DashboardHeader title="Analytics" subtitle="Activity at a glance" />
-        <DashboardContent>
-          <DashboardGrid columns={3}>
-            {[1, 2, 3, 4, 5, 6].map((i) => (
-              <StatCardSkeleton key={i} />
-            ))}
-          </DashboardGrid>
-        </DashboardContent>
-      </>
-    );
-  }
+  const queryClient = new QueryClient();
 
-  const currency = data.capabilities.currency ?? "INR";
-  const learners = countByRole(data.members.byRole, "LEARNER");
-  const experts = countByRole(data.members.byRole, "EXPERT");
-
-  // Earnings "paid" cell: sum the orgShare of PAID rows, net of refunds.
-  const paidEarnings =
-    data.earnings?.find((e) => e.status === "PAID")?.orgSharePaise ?? 0;
-  const refundedEarnings = (data.earnings ?? []).reduce(
-    (sum, e) => sum + e.refundedPaise,
-    0,
-  );
+  // queryKey MUST match AnalyticsPageClient's useQuery
+  // (["org-analytics", orgId]) or hydration won't apply.
+  await Promise.allSettled([
+    queryClient.prefetchQuery({
+      queryKey: ["org-analytics", orgId],
+      queryFn: () => getOrgAnalytics(orgId),
+    }),
+  ]);
 
   return (
-    <>
-      <DashboardHeader title="Analytics" subtitle="Activity at a glance" />
-      <DashboardContent>
-        <DashboardGrid columns={3}>
-          <StatCard
-            title="Members"
-            value={data.members.total}
-            subtitle={`${data.members.active} active`}
-            icon={Users}
-            variant="info"
-          />
-          {data.capabilities.canSponsor && (
-            <StatCard
-              title="Learners"
-              value={learners}
-              subtitle={
-                data.programs.activeAssignments > 0
-                  ? `${data.programs.activeAssignments} active assignments`
-                  : "No active assignments"
-              }
-              icon={UserCheck}
-            />
-          )}
-          {data.capabilities.canHost && (
-            <StatCard title="Experts" value={experts} icon={UserCheck} />
-          )}
-          <StatCard
-            title="Active programs"
-            value={data.programs.active}
-            subtitle={`${data.programs.total} total`}
-            icon={Briefcase}
-          />
-          {data.wallet && (
-            <StatCard
-              title="Wallet balance"
-              value={formatCurrencyAmount(data.wallet.balancePaise, currency)}
-              icon={Wallet}
-              variant="success"
-            />
-          )}
-          {data.invoices && (
-            <>
-              <StatCard
-                title="Outstanding invoices"
-                value={data.invoices.outstandingCount}
-                subtitle={formatCurrencyAmount(
-                  data.invoices.outstandingPaise,
-                  currency,
-                )}
-                icon={FileText}
-                variant={data.invoices.pastDueCount > 0 ? "warning" : "info"}
-              />
-              <StatCard
-                title="Paid (last 30 days)"
-                value={data.invoices.paidLast30dCount}
-                subtitle={formatCurrencyAmount(
-                  data.invoices.paidLast30dPaise,
-                  currency,
-                )}
-                icon={TrendingUp}
-                variant="success"
-              />
-              {data.invoices.pastDueCount > 0 && (
-                <StatCard
-                  title="Past-due invoices"
-                  value={data.invoices.pastDueCount}
-                  icon={AlertCircle}
-                  variant="warning"
-                />
-              )}
-            </>
-          )}
-          {data.earnings && data.earnings.length > 0 && (
-            <StatCard
-              title="Earnings — paid"
-              value={formatCurrencyAmount(paidEarnings, currency)}
-              subtitle={
-                refundedEarnings > 0
-                  ? `${formatCurrencyAmount(refundedEarnings, currency)} refunded`
-                  : undefined
-              }
-              icon={Wallet}
-              variant="success"
-            />
-          )}
-        </DashboardGrid>
-      </DashboardContent>
-    </>
+    <HydrationBoundary state={dehydrate(queryClient)}>
+      <AnalyticsPageClient orgId={orgId} />
+    </HydrationBoundary>
   );
 }
